Rename namespace to Pheanstalk
